lucy/pyproject.toml
[project]
name = "lucy01"
version = "1.1.0"
description = "CLI companion for CP."
readme = "README.md"
dependencies = [
    "bs4",
    "click",
    "lxml",
    "python-dotenv",
    "requests",
    "selenium",
]
requires-python = ">=3.10"
authors = [{ name = "Mehul Todi", email = "mehul116.dev@gmail.com" }]
keywords = [
    "cp",
    "cli",
    "companion",
    "atcoder",
    "codeforces",
    "tester",
    "snippets",
    "lucy"
]
classifiers = [
    "Development Status :: 4 - Beta",

    "Programming Language :: Python :: 3.10",

    "Operating System :: POSIX :: Linux",

    "Intended Audience :: Education",

    "Topic :: Education",
]
# license = {text = "MIT License"}

[project.urls]
# Homepage = "https://example.com"
Documentation = "https://lucy01.readthedocs.io/en/latest/"
Repository = "https://github.com/kid-116/Lucy"
Issues = "https://github.com/kid-116/Lucy/issues"
# Changelog = "https://github.com/me/spam/blob/master/CHANGELOG.md"

[project.optional-dependencies]
lint = ["mypy", "pylint", "pylint[spelling]", "yapf"]
docs = ["mkdocs", "md-click@git+https://github.com/kid-116/md-click.git@support-arguments"]
test = ["pytest", "pytest-datadir"]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project.scripts]
lucy = "lucy.main:lucy"

[tool.mypy]
non_interactive = true
install_types = true
strict = true
exclude = "build|venv"

[tool.pylint]
ignore = ["venv", "build"]
[tool.pylint.'MESSAGES CONTROL']
disable = "missing-module-docstring,missing-function-docstring,missing-class-docstring,missing-function-docstring"
[tool.pylint.'SPELLING']
spelling-dict = 'en_US'
spelling-private-dict-file = 'private-dictionary.txt'

[tool.yapf]
based_on_style = "google"
column_limit = 100
[tool.yapfignore]
ignore_patterns = ["venv/**"]

[tool.setuptools]
# packages = ["lucy"]
include-package-data = true

[tool.setuptools.package-data]
'lucy' = ['data/template_cpp.cpp']
